Refactor Num's isNumeric() and val()
I moved the regex for fraction strings and the regex for mixed numbers to a
class constant. I needed it multiple places, and created separate methods to
test and parse the strings seemed like a waste. I updated val() to more closely
match PHP's native intval() and floatval() methods, particularly how they
treat arrays and objects. As a result, I had to refactor the isNumeric() method
to not rely on the val() method.