Fix logic errors in Arr
Fix the logic in Arr::inArray() that tests to see if the needle starts or ends
with the wildcard character. It was using the inArray()'s haystack, an array,
instead of the inArray()'s needle argument, a string. Also, fix the default
value of the isEmpty() return value. Default to true instead of false.